Add partial support for nonstandard Injectors
Non-void injectors are not really supported due to them being rather painful
to implement. If anyone wants to, they can do it - but it'd require stack
manipulation that I'm not really up to dealing myself.

Especially with the CallbackInfo chaining that I have in my mind such
stack manipulation goes nowhere.