Injection point selectors select instructions, not labels
I am not well sure why I initially implemented the injection selection behaviour
based on LabelNodes, but this behaviour was altered to now directly match the appropriate
AbstractInsnNodes. The matched instruction may NOT be a virtual instruction (e.g.
LabelNode or FrameNode).

The new behaviour is not problematic on empty methods as the RETURN opcode should exist
for every method. Furthermore the integration tests show no signs of any regressions.