[CI-SKIP] Introduce tests to validate the unviability of Intrinsic
We will most likely be using `@Unique` in the meantime.

From one comment of mine on sponge's discord:
Nevermind, that wouldn't work as `@Intrinisc` requires a prefix (my assumption
is that it has a hardcoded check for `$` - which is why the mixin documentation
for the prefixes is a bit fuzzy or even contradicting itself as that hardcoded
check probably wasn't intended at first but was declared WAI later on).
However, as `@Intrinsic` does not have a prefix element in of itself, it relies
on other annotations to provide the prefix. So in the end `@Intrinsic`'s only
purpose is to soft-implement an interface and what I'm looking after
(hard-implementing an interface) should be done via `@Unique`.