0.5.0 breaking changes
The bulk of this change is moving most classes which weren't
explicitly API nor implementation-specific for the api package.
The only exception is SimpleRemapper which was moved to the implementation
package.

Furthermore, the logging interface now has a debug logging method,
although it is not being utilized as of yet.