Prepare publication to OSSRH: Signing
I don't actually know how the current CI setup will react to these changes,
which is why I make this commit without bundling it without any other changes
in order to be able to test their effects and if needed readjust the CI scripts
used to compile micromixin.

Also this is among one of the last steps needed beware we can start publishing
releases to OSSRH in a semi-regular interval. But things remain to be seen whether
that will actually happen (especially since I still have to claim a groupId
for this)