Fix Redirect not properly pushing this down the operand stack
It would occasionally cause category 2 computational types to be cut off, which
is behaviour not tolerated by the Verifier. Now, Redirect uses almost the same
logic as ModifyArg, although ModifyArg's stack head moving logic had to be slightly
adapted in order to accomodate scenarios where nothing needs to be pushed down the
stack. This is primiarily an issue as for Redirect you can be quite sure there is
background (unless you'd want to compute it but that might be a bit on the
expensive side of things for little benefit)