Deprecate the Objects class for external consumption
Far too often this class would get accidentally be used by projects
using this class as a transitive dependency. Due to the fact that imports
are quite frequently well hidden by the IDE this fact usually flies under
the radar. By enforcing a deprecation we hope to increase awareness
of the internal properties of this class.

However while this is especially the case with the Objects class,
all other classes in the internal package should be deemed as internal API
and is not to be used by dependents - whether accidentally or not.