Inclusive guard checking
This set of changes is a followup to #4 and modifies the behavior of @At matching,
that way the loop ends on the guard instruction, including guard.
Previous behavior specified guard as the instruction after matched which
required extra processing (like skipping frames or labels) as seen in
SlicedInjectionPointSelector.getAfterSelected.

It was still flawed regardless, as when the matching instruction is the
last instruction in MethodNode.instructions, it will end up on null,
which means it will throw an IllegalStateException about exhausted instruction list