Fix malformed SMAP attributes - again
Some of these bugs have existed since the inception of the SMAP integration,
others are newly introduced. Regardless, regression tests are in place now to
hopefully stem new bugs from arising in the future. Whether that labour bears
fruit remains to be seen, as it isn't exactly something we can evaluate ahead
of time without guessing - can we?