Copy LVT into handler methods
This further improves debug support under eclipse.
There remain a few rough edges, but those can be polished too with enough effort.

Further, this commit also did a slight refractor inside CodeCopyUtil, cutting down
the amount of arguments that need to be passed in some circumstances. However, the
amount of arguments within the methods of that class remain too high and the order
of the arguments are possibly counter-intuitive with the current design
philosophies I employ under the stianloader umbrella.