Introduce micromixin-testing-suite-neo; Deprecate micromixin-runtime
Currently micromixin-testing-suite-neo is a skeleton of what it could be,
though I am fairly certain that it will grow over time.

This is also the first full-scale JPMS of SLL. Launcher-micromixin
felt easier on the bones, but that might be caused by launcher-sponge
having done all the hard work already. Regardless, launcher-micromixin
did show that there was a serious flaw with how micromixin-runtime was
handled, so we have decided to deprecate the package. It will be removed
shortly (within the 0.8.0 cycle), meaning that newer artifacts will not be
published.

Also unlike the old testing suite (which will be kept in this repository for
a while at least), the output of the testing suite can be viewed in the
jenkins log. Well at least hopefully - it's a bit more difficult to test
your jenkins config locally so we'll have to hope that I got it right on
the first try.